Updated Portfolio Page
Moved articles around on the main html index file. Added in my new project as the main focused post and adjusted the scroll feature to move to the main rather than the header. Added the html file for the new project and moved the board game post to its own article section classified like a main post to prevent disjointed appearance.